last update

+++ app/src/main/java/com/example/covid19_vaccine/data/domain/StatisticRecyclerView.kt
@@ -1,8 +1,8 @@
 package com.example.covid19_vaccine.data.domain
 
-data class StatisticRecyclerView (
-    val countryName : String,
-    val countryVaccinePer: Int,
-    val countryVaccineTotal : Double
+data class StatisticRecyclerView(
+    val countryName: String,
+    val countryVaccinePer: Double,
+    val countryVaccineTotal: Double
 
 )
